Refactor factory use
Creates factory subclasses for each factory and use global variables
to hold the factory instantiation rather than static fields in the
class the factory is for.

Also separates the child classes added in this PR into individual
files. `get_wf_tasks()` had to be separated from `Tasks` and
`create_wf_task()` to avoid circular imports.

Refs: #1577
